refactor: Move not shared code from SharedUtils to MainVM
