+++ app/src/main/java/com/infomaniak/mail/utils/SharedUtils.kt
@@ -32,7 +32,6 @@ import com.infomaniak.mail.data.cache.mailboxContent.RefreshController.RefreshMo
 import com.infomaniak.mail.data.cache.mailboxContent.ThreadController
 import com.infomaniak.mail.data.cache.mailboxInfo.MailboxController
 import com.infomaniak.mail.data.models.mailbox.Mailbox
-import com.infomaniak.mail.data.models.mailbox.SendersRestrictions
 import com.infomaniak.mail.data.models.message.Message
 import com.infomaniak.mail.data.models.thread.Thread
 import com.infomaniak.mail.ui.main.settings.SettingRadioGroupView
@@ -131,40 +130,6 @@ class SharedUtils @Inject constructor(
                 realm = realm,
                 callbacks = if (folderId == currentFolderId) callbacks else null,
             )
-        }
-    }
-
-    suspend fun unblockSender(mailbox: Mailbox, emailToUnblock: String) {
-        with(ApiRepository.getSendersRestrictions(mailbox.hostingId, mailbox.mailboxName)) {
-            if (isSuccess()) {
-                val updatedSendersRestrictions = SendersRestrictions().apply {
-                    authorizedSenders = data!!.authorizedSenders
-                    blockedSenders = data!!.blockedSenders.apply {
-                        removeIf { it.email == emailToUnblock }
-                    }
-                }
-                updateBlockedSenders(mailbox, updatedSendersRestrictions)
-            }
-        }
-    }
-
-    private suspend fun updateBlockedSenders(mailbox: Mailbox, updatedSendersRestrictions: SendersRestrictions) {
-        with(ApiRepository.updateBlockedSenders(mailbox.hostingId, mailbox.mailboxName, updatedSendersRestrictions)) {
-            if (isSuccess()) {
-                mailboxController.updateMailbox(mailbox.objectId) {
-                    it.sendersRestrictions = updatedSendersRestrictions
-                }
-            }
-        }
-    }
-
-    suspend fun updateSendersRestrictions(mailbox: Mailbox) {
-        with(ApiRepository.getSendersRestrictions(mailbox.hostingId, mailbox.mailboxName)) {
-            if (isSuccess()) {
-                mailboxController.updateMailbox(mailbox.objectId) {
-                    it.sendersRestrictions = data
-                }
-            }
         }
     }
